GCC/Clang fixes, pedantic GCC/Clang warnings and added compression level control. - GCC/Clang compile fixes. - Enabled GCC/Clang pedantic warnings. - Bug fixes highlighted by additional warnings. - Zero warnings pass. - Added ServerSettings CompressionLevel control for CollatedPacket. - Reduced default CollatedPacket compression to improve performance.

+++ 3es-core/3esconnectionmonitor.h
@@ -86,7 +86,7 @@ namespace tes
     /// This may be TCP specific.
     ///
     /// @return The port connections are being monitored on.
-    virtual int port() const = 0;
+    virtual unsigned short port() const = 0;
 
     /// Starts the monitor listening in the specified mode.
     ///
